Fixes profile issue: 3.2 Grammar Learning Disabling Parameters - In particular, the AT(xsi:type) productions that would be inserted in grammars that would be instantiated after the maximumNumberOfBuiltInElementGrammars threshold are not counted.
It encodes the following XML.

<A><B/><B/><B/></A>

maxBuiltinGr ="1"
maxBuiltinProd="2"

The expectation is that xsi:type productions for <B/> never gets counted
because by the time first <B/> appears, the limit for built-in grammars
(it is 1 in this case) has already been reached. This makes for a room
for the second <B> to be learned by A's ElementContent grammar.