Handle IllegalArgumentException when decoding
IllegalArgumentException can happen when decoding a file, specifically
when attempting to decode a schema aware file without using a schema
aware decoder. This wasn't being cleanly handled before and as a result
the exception was bubbling up into SAX and getting printed out to stderr
regardless if the exception was being caught above SAX.

This commit catches the IllegalArgumentException and converts it to a
SAXException, which is easier to deal with and can be handled in SAX by
setting up and appropriate SAXErrorHandler.

https://github.com/EXIficient/exificient/issues/33